Split the Home file
