+++ packages/toolpad-app/src/toolpad/Home/index.tsx
@@ -17,7 +17,6 @@ import {
   MenuItem,
   Skeleton,
   Stack,
-  Snackbar,
   Table,
   TableBody,
   TableCell,
@@ -31,30 +30,26 @@ import {
 } from '@mui/material';
 import { LoadingButton } from '@mui/lab';
 import IconButton from '@mui/material/IconButton';
-import CloseIcon from '@mui/icons-material/Close';
-import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
 import DriveFileRenameOutlineIcon from '@mui/icons-material/DriveFileRenameOutline';
 import ViewListIcon from '@mui/icons-material/ViewList';
 import GridViewIcon from '@mui/icons-material/GridView';
 import DeleteIcon from '@mui/icons-material/Delete';
-import client from '../api';
-import DialogForm from '../components/DialogForm';
-import type { Deployment } from '../../prisma/generated/client';
-import useLatest from '../utils/useLatest';
-import ToolpadShell from './ToolpadShell';
-import getReadableDuration from '../utils/readableDuration';
-import EditableText from '../components/EditableText';
-import type { AppMeta } from '../server/data';
-import useMenu from '../utils/useMenu';
-import useLocalStorageState from '../utils/useLocalStorageState';
-import ErrorAlert from './AppEditor/PageEditor/ErrorAlert';
-import { ConfirmDialog } from '../components/SystemDialogs';
-import config from '../config';
-import { errorFrom } from '../utils/errors';
-import { DOCUMENTATION_URL, TOOLPAD_TARGET_CLOUD } from '../constants';
-
-const CURRENT_RELEASE_VERSION = `v${process.env.TOOLPAD_VERSION}`;
+import client from '../../api';
+import DialogForm from '../../components/DialogForm';
+import type { Deployment } from '../../../prisma/generated/client';
+import useLatest from '../../utils/useLatest';
+import ToolpadShell from '../ToolpadShell';
+import UpdateBanner from './UpdateBanner';
+import getReadableDuration from '../../utils/readableDuration';
+import EditableText from '../../components/EditableText';
+import type { AppMeta } from '../../server/data';
+import useMenu from '../../utils/useMenu';
+import useLocalStorageState from '../../utils/useLocalStorageState';
+import ErrorAlert from '../AppEditor/PageEditor/ErrorAlert';
+import { ConfirmDialog } from '../../components/SystemDialogs';
+import config from '../../config';
+import { errorFrom } from '../../utils/errors';
 
 export interface CreateAppDialogProps {
   open: boolean;
@@ -473,72 +468,6 @@ function AppsListView({ loading, apps, activeDeploymentsByApp, setDeletedApp }: 
       </TableBody>
     </Table>
   );
-}
-
-function UpdateBanner() {
-  const { data: latestRelease } = client.useQuery('getLatestToolpadRelease', [], {
-    staleTime: 1000 * 60 * 10,
-    enabled: process.env.TOOLPAD_TARGET !== TOOLPAD_TARGET_CLOUD,
-  });
-
-  const [dismissedVersion, setDismissedVersion] = useLocalStorageState<string | null>(
-    'update-banner-dismissed-version',
-    null,
-  );
-
-  const handleDismissClick = React.useCallback(
-    (event: React.SyntheticEvent | Event, reason?: string) => {
-      if (reason === 'clickaway' || reason === 'escapeKeyDown') {
-        return;
-      }
-      setDismissedVersion(CURRENT_RELEASE_VERSION);
-    },
-    [setDismissedVersion],
-  );
-
-  const hideBanner =
-    (latestRelease && latestRelease.tag === CURRENT_RELEASE_VERSION) ||
-    dismissedVersion === CURRENT_RELEASE_VERSION;
-
-  return latestRelease ? (
-    <Snackbar
-      open={!hideBanner}
-      onClose={handleDismissClick}
-      anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-      message={
-        <React.Fragment>
-          A new version <strong>{latestRelease.tag}</strong> of Toolpad is available.
-        </React.Fragment>
-      }
-      action={
-        <Stack direction="row" sx={{ gap: 2 }}>
-          <Button
-            aria-label="update"
-            color="inherit"
-            endIcon={<OpenInNewIcon fontSize="inherit" />}
-            component="a"
-            target="_blank"
-            href={DOCUMENTATION_URL}
-          >
-            Update
-          </Button>
-          <Button
-            aria-label="view changelog"
-            color="inherit"
-            endIcon={<OpenInNewIcon fontSize="inherit" />}
-            component="a"
-            target="_blank"
-            href={latestRelease.url}
-          >
-            View changelog
-          </Button>
-          <IconButton aria-label="close" color="inherit" size="small" onClick={handleDismissClick}>
-            <CloseIcon fontSize="inherit" />
-          </IconButton>
-        </Stack>
-      }
-    />
-  ) : null;
 }
 
 export default function Home() {
